requirements.txt
asgiref==3.8.1
Django==4.2
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.0.0
inflection==0.5.1
mccabe==0.7.0
packaging==24.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1